+++ CHANGELOG.md
@@ -12,6 +12,11 @@ This project adheres to Semantic Versioning.
 
 * `window::set_screen_saver_enabled` and `window::is_screen_saver_enabled` have been added, allowing you to control whether the user's screensaver can activate while playing your game. ([@sumibi-yakitori](https://github.com/sumibi-yakitori) in [#216](https://github.com/17cupsofcoffee/tetra/pull/216))
 
+### Changed
+
+* The game loop now calls `std::thread::sleep` instead of `std::thread::yield_now` at the end of each iteration.
+    * This provides some basic frame-limiting when running without vsync, and prevents CPU usage skyrocketing on some systems.
+
 ## [0.5.5] - 2020-12-03
 
 ### Added